Support gradients for text
To do this we need to pain text in a two pass process. First we compute the layout for the
glyph run. During this we keep track of the complete text bounds. Then in a separate pass we do
the actual painting. The text bounds are passed to the gradient paint.

Fixes #15