Ensure forward referenced nodes being used by href have been built
Previously a node template (e.g. for gradients) specified by 'href' could
be in an invalid state because ParsedElement#build wasn't called yet.
We now track the build status of such elements and request it to be build
if it is being referenced. If a cyclic reference is specified we issue a
warning and continue building. This may result in an error further down
(which is fine as the svg is malformed in this case).
We should probably be more conformal to the spec regarding this but the
current behaviour will do for now.